feat: move all sequence_player code back to sequence, add Slide and SlidePlayer
